requirements.txt
markdown
xhtml2pdf
streamlit
google-generativeai